Update rough-draft
he Right Mindset: Reflections on Carol Dweck's Motivation Research & The Power of "Yet"
The concept of adopting a "non-fixed" or fluid, adaptive mindset rather than a "fixed", or rigid frame of mind, in order to more efficiently acquire new knowledge.  This concept has been the focus of research for Stanford Researcher, Carol Dweck, aka "the mindset lady". 
Through years of careful observation and scientific empirical testing, Carol Dweck has formulated a system, or rather a way of thinking based on this research for sustained cognitive improvement and improved academic learning.
This new way of approaching the subject of academic motivation and learning is simply called "Yet" thinking.
The "Yet" concept came about while Carol Dweck was observing a schoolroom, where, instead of students receiving an "F", the teacher would write "Not Yet" which, as the research has shown, psychologically transmits to the student the idea that learning is a process, which, when performed of confronting unfamiliar territory, experimenting with ways of illuminating the situation, struggling with the unfamiliar, but all the while approaching it not with fear or pre-emptive anxiety, but with the expectant joy of confronting an exciting challenge that could turn into a valuable learning experience.  In the past this has been  It has been shown that this  improves motivation and academic performance, especially in the young minds of children. Instead of treating learning as an exercise in problem solving, the "Yet" concept approaches learning as a process or a journey.  
As an aside, it is important to point out that the findings of this research and the subsequent defining of the concept of "yet" thinking is not only effective for children, but people of all ages.  

The Stanford researcher's subjects were elementary school children from failing school systems, some substandard schools, such as a school in the Bronx, NY, and groups who have unfortunately acquired the socio-academic stigma of being "under achievers" or inherently cut off from higher learning by an expectation of failure. The "yet" concept is used to describe a certain way of positive reinforcement devoid of finite labels such as "Fail". Instead, children were told that if they keep practicing what they are learning. math skills, for instance, their brains, although initially terrified of the challenges, begins to actually develop new pathways of growth.